Revert "install node.js on running bvm link"
This reverts commit bc06727a44e978c273e37d11f96052267a828602.
