+++ workspace.jsonc
@@ -58,7 +58,7 @@
         "@types/tar": "4.0.4",
         "@types/tar-fs": "2.0.0",
         "@types/yargs": "17.0.0",
-        "@zkochan/cmd-shim": "5.3.0",
+        "@zkochan/cmd-shim": "5.2.2",
         "bin-links": "2.2.1",
         "chalk": "4.1.0",
         "cli-progress": "3.10.0",
